✨ Support Sentry V4
