+++ composer.json
@@ -37,14 +37,14 @@
     "require": {
         "php": "^7.4 || ^8.0",
         "monolog/monolog": "^1.6 || ^2.0 || ^3.0",
-        "sentry/sentry": "^3.1"
+        "sentry/sentry": "^3.1 || ^4.0"
     },
     "require-dev": {
         "coduo/php-matcher": "^6.0.8",
-        "friendsofphp/php-cs-fixer": "^3.16",
+        "friendsofphp/php-cs-fixer": "3.51.0",
         "jangregor/phpstan-prophecy": "1.0.0",
-        "phpstan/phpstan": "1.10.14",
-        "phpstan/phpstan-phpunit": "1.3.11",
+        "phpstan/phpstan": "1.10.59",
+        "phpstan/phpstan-phpunit": "1.3.16",
         "phpunit/phpunit": "^9.6.7",
         "symfony/phpunit-bridge": "^6.2.10",
         "symfony/var-dumper": ">4.3"
